完善Post单元测试，Post内容、Cookie全部通过
