+++ src/Senparc.Weixin.MP/Senparc.WeixinTests/Utilities/HttpUtility/RequestUtilityTests.cs
@@ -22,7 +22,9 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Senparc.Weixin.HttpUtility;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
+using System.Net;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -40,5 +42,51 @@ namespace Senparc.Weixin.HttpUtility.Tests
             //清除
             RequestUtility.RemoveHttpProxy();
         }
+
+
+        [TestMethod]
+        public void PostTest()
+        {
+            var data = "Jeffrey";
+            Stream stream = new MemoryStream();
+            var bytes = Encoding.UTF8.GetBytes(data);
+            stream.Write(bytes, 0, bytes.Length);
+            stream.Seek(0, SeekOrigin.Begin);
+
+            var cookieContainer = new CookieContainer();
+            var url = "http://localhost:65395/ForTest/PostTest";//使用.NET 4.5的Sample
+            var result = Senparc.Weixin.HttpUtility.RequestUtility.HttpPost(url,
+                cookieContainer, stream, useAjax: true);
+
+            Console.WriteLine(result);
+
+            Assert.IsNotNull(result);
+        }
+
+
+        [TestMethod]
+        public void SenparcHttpResponseTest()
+        {
+            var data = "Jeffrey";
+            Stream stream = new MemoryStream();
+            var bytes = Encoding.UTF8.GetBytes(data);
+            stream.Write(bytes, 0, bytes.Length);
+            stream.Seek(0, SeekOrigin.Begin);
+
+            var cookieContainer = new CookieContainer();
+            var url = "http://localhost:65395/ForTest/PostTest";//使用.NET 4.5的Sample
+            var result = Senparc.Weixin.HttpUtility.RequestUtility.HttpResponsePost(url,
+                cookieContainer, stream, useAjax: true);
+
+            Assert.IsNotNull(result);
+
+            var resultString = result.Result.Content.ReadAsStringAsync().GetAwaiter().GetResult();
+            Console.WriteLine("resultString:{0}", resultString);
+
+            var cookie = cookieContainer.GetCookies(new Uri("http://localhost:65395"));
+            Console.WriteLine("TestCookie：{0}", cookie["TestCookie"]);
+        }
+
+
     }
 }
